character tests

+++ app/Http/Controllers/CharacterController.php
@@ -14,6 +14,7 @@ use Illuminate\Http\JsonResponse;
  */
 class CharacterController extends Controller
 {
+
     /**
      * Display a listing of the resource.
      *
@@ -22,7 +23,7 @@ class CharacterController extends Controller
     public function index()
     {
         //todo character search endpoint
-        return response()->json(['error'=>'incomplete endpoint'], JsonResponse::HTTP_IM_USED);
+        return response()->json(['error' => 'incomplete endpoint']);
     }
 
     /**
@@ -33,20 +34,21 @@ class CharacterController extends Controller
     public function create()
     {
         //todo show form for creating a character
-        return response()->json(['error'=>'incomplete endpoint']);
+        return response()->json(['error' => 'incomplete endpoint']);
     }
 
     /**
      * Store a newly created resource in storage.
      *
-     * @param  CreateCharacterRequest  $request
+     * @param  CreateCharacterRequest $request
+     *
      * @return \Illuminate\Http\Response
      */
     public function store(CreateCharacterRequest $request)
     {
         $user = \Auth::user();
 
-        if (!$user){
+        if ( ! $user) {
             //todo redirect
         }
 
@@ -62,23 +64,28 @@ class CharacterController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  string  $id
+     * @param  string $id
+     *
      * @return \Illuminate\Http\Response
      */
     public function show(string $id)
     {
-        $character = Character::whereId($id);
+        $character = Character::whereId($id)->first();
+
         //todo create view to see character
-        return response()->json(['error'=>'incomplete endpoint', 'character'=>$character->toArray()], JsonResponse::HTTP_IM_USED);
+        return response()->json(['error' => 'incomplete endpoint', 'character' => $character->toArray()]);
     }
 
     /**
+     * Return a Chararacter
+     *
      * @param string $id
+     *
      * @return \Illuminate\Http\Response
      */
     public function get(string $id)
     {
-        $character = Character::whereId($id);
+        $character = Character::whereId($id)->first();
 
         return response()->json($character->toArray());
     }
@@ -86,26 +93,49 @@ class CharacterController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  string  $id
+     * @param  string $id
+     *
      * @return \Illuminate\Http\Response
      */
     public function edit(string $id)
     {
-        $character = Character::whereId($id);
+        $user = \Auth::user();
+
+        if ( ! $user) {
+            //todo redirect
+        }
+
+        $character = Character::whereId($id)->first();
+
+        if ($user->id != $character->owner_id) {
+            return response()->json(['error' => 'not allowed'], JsonResponse::HTTP_UNAUTHORIZED);
+        }
+
         //todo create a view to edit a character
-        return response()->json(['error'=>'incomplete endpoint', 'character'=>$character->toArray()], JsonResponse::HTTP_IM_USED);
+        return response()->json(['error' => 'incomplete endpoint', 'character' => $character->toArray()]);
     }
 
     /**
      * Update the specified resource in storage.
      *
-     * @param  UpdateCharacterRequest  $request
-     * @param  string  $id
+     * @param  UpdateCharacterRequest $request
+     * @param  string $id
+     *
      * @return \Illuminate\Http\Response
      */
     public function update(UpdateCharacterRequest $request, string $id)
     {
-        $character = Character::whereId($id);
+        $user = \Auth::user();
+
+        if ( ! $user) {
+            //todo redirect
+        }
+
+        $character = Character::whereId($id)->first();
+
+        if ($user->id != $character->owner_id) {
+            return response()->json([], JsonResponse::HTTP_UNAUTHORIZED);
+        }
 
         $character->update($request->toArray());
 
@@ -117,12 +147,23 @@ class CharacterController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  string  $id
+     * @param  string $id
+     *
      * @return \Illuminate\Http\Response
      */
     public function destroy(string $id)
     {
-        $character = Character::whereId($id);
+        $user = \Auth::user();
+
+        if ( ! $user) {
+            //todo redirect
+        }
+
+        $character = Character::whereId($id)->first();
+
+        if ($user->id != $character->owner_id) {
+            return response()->json([], JsonResponse::HTTP_UNAUTHORIZED);
+        }
 
         $character->delete();
